wip: add edit deck route

+++ src/app/api/v1/decks/[id]/route.ts
@@ -1,8 +1,11 @@
 import { NextResponse } from "next/server";
 import { deleteDeck, getDeckById, updateDeck } from "../../repositories/deck";
+import { getServerSession } from "next-auth";
+import { STATUS_CODES } from "http";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 export const GET = async (req: Request, context: { params: { id: string } }) => {
-    const { id } = context.params
+    const { id } = await context.params
 
     try {
         const data = await getDeckById(id)
@@ -16,11 +19,22 @@ export const GET = async (req: Request, context: { params: { id: string } }) => 
 }
 
 export const PUT = async (req: Request, context: { params: { id: string } }) => {
+    const data = await getServerSession(authOptions);
     const { deck } = await req.json()
-    const { id } = context.params
+    const { id } = await context.params
+
+    if (!data)
+        return new NextResponse(JSON.stringify({ success: false, message: STATUS_CODES[401] }),
+            { status: 401, headers: { 'Content-Type': 'application/json' } });
+
+    const compDeck = await getDeckById(id)
+
+    if (data.user.id != compDeck[0].decks.userId)
+        return new NextResponse(JSON.stringify({ success: false, message: STATUS_CODES[403] }),
+            { status: 403, headers: { 'Content-Type': 'application/json' } });
 
     try {
-        updateDeck(id, deck)
+        await updateDeck(id, deck)
 
         return new NextResponse(JSON.stringify({ success: true }), { status: 200, headers: { 'Content-Type': 'application/json' } });
     } catch (err) {
@@ -31,10 +45,25 @@ export const PUT = async (req: Request, context: { params: { id: string } }) => 
 }
 
 export const DELETE = async (_: Request, context: { params: { id: string } }) => {
-    const { id } = context.params
+    const data = await getServerSession(authOptions);
+
+    const { id } = await context.params
+
+    if (!data)
+        return new NextResponse(JSON.stringify({ success: false, message: STATUS_CODES[401] }),
+            { status: 401, headers: { 'Content-Type': 'application/json' } });
+
+    const compDeck = await getDeckById(id)
+
+    console.log(compDeck)
+    console.log(data)
+
+    if (data.user.id != compDeck[0].decks.userId)
+        return new NextResponse(JSON.stringify({ success: false, message: STATUS_CODES[403] }),
+            { status: 403, headers: { 'Content-Type': 'application/json' } });
 
     try {
-        deleteDeck(id)
+        await deleteDeck(id)
 
         return new NextResponse(JSON.stringify({ success: true }), { status: 200, headers: { 'Content-Type': 'application/json' } });
     } catch (err) {
